# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/authrequired/admin/admin": {
            "delete": {
                "description": "删除指定管理员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "删除指定管理员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "需要查询的邮箱",
                        "name": "queryemail",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "需要查询ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "需要查询用户名",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            },
            "patch": {
                "description": "修改指定管理员信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "修改指定管理员信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "需要查询的学生邮箱",
                        "name": "queryemail",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "需要查询学生ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "需要查询学生用户名",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "修改邮箱",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "修改真实姓名",
                        "name": "realName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "修改学号",
                        "name": "userSchoollD",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "修改学校代码",
                        "name": "schoolCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "修改班级",
                        "name": "class",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "修改性别",
                        "name": "sex",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/authrequired/admin/admin/new": {
            "post": {
                "description": "注册管理员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "注册管理员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新用户用户名",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "新用户邮箱",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新用户真实姓名",
                        "name": "realName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新用户性别",
                        "name": "sex",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/authrequired/admin/admin/password": {
            "patch": {
                "description": "修改指定管理员密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "修改指定管理员密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "需要查询的学生邮箱",
                        "name": "queryemail",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "需要查询学生ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "需要查询学生用户名",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
...
        "/authrequired/admin/student/password": {
            "patch": {
                "description": "修改指定学生密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "修改指定学生密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "需要查询的学生邮箱",
                        "name": "queryemail",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "需要查询学生ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "需要查询学生用户名",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "新密码",
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errhandle.ErrCode": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14
            ],
            "x-enum-varnames": [
                "NoError",
                "InnerError",
                "TokenError",
                "ParamsError",
                "UserExists",
                "UserNonExists",
                "PasswordInvalid",
                "PermissionDenied",
                "ClassError",
                "SexError",
                "NameExists",
                "EmailFormatError",
                "RealNameFormatError",
                "SchoolError",
                "PasswordTooShort"
            ]
        },
        "model.CommonResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/errhandle.ErrCode"
                }
            }
        },
...
    }
}